+++ system.go
@@ -32,9 +32,10 @@ func versionError(got string) error {
 
 func (c *Client) versionCheck(ctx context.Context, depth int32) error {
 
+	c.compatibilityErr = nil
 	sv := systemVersion{}
 	if err := c.jsonGet(ctx, systeminfoAPI, &sv, depth); err != nil {
-		return err
+		return fmt.Errorf("system info error %d (%w)", depth, err)
 	}
 
 	if !sv.Ready {